Feature 35 document language update lambda

+++ src/functions/create-session/handler.ts
@@ -1,7 +1,7 @@
 import type { ValidatedEventAPIGatewayProxyEvent } from "@libs/api-gateway";
 import { middyfy } from "@libs/lambda";
 import { authenticationService } from "src/database/services";
-import { generateToken, parseBasicAuth, parseBearerAuth } from "@libs/auth-utils";
+import { generateToken, parseBasicAuth, parseBearerAuth, returnForbidden, returnUnauthorized } from "@libs/auth-utils";
 import { generateExpiryTimestamp, validateTimestamp } from "@libs/date-utils";
 import { v4 as uuid } from "uuid";
 import { getElastic } from "src/elastic";
@@ -18,37 +18,25 @@ const createAuthenticationSession: ValidatedEventAPIGatewayProxyEvent<any> = asy
   const isBasicAuth = authHeader?.toLowerCase()?.startsWith("basic");
   const isBearerAuth = authHeader?.toLowerCase()?.startsWith("bearer");
 
-  if (!isBearerAuth && !isBasicAuth) {
-    return {
-      statusCode: 401,
-      body: "Unauthorized"
-    };
+  if (!(isBearerAuth || isBasicAuth)) {
+    return returnUnauthorized();
   }
 
   if (isBearerAuth) {
     const token = parseBearerAuth(authHeader!);
     if (!token) {
-      return {
-        statusCode: 401,
-        body: "Unauthorized"
-      };
+      return returnUnauthorized();
     }
 
     const foundSession = await authenticationService.findSession(token);
     if (!foundSession) {
-      return {
-        statusCode: 401,
-        body: "Unauthorized"
-      };
+      return returnUnauthorized()
     }
 
     if (!validateTimestamp(foundSession.expiresAt)) {
       await authenticationService.deleteSession(foundSession.token);
 
-      return {
-        statusCode: 401,
-        body: "Unauthorized"
-      }
+      return returnUnauthorized()
     }
 
     const tokenExpiry: number = generateExpiryTimestamp();
@@ -72,18 +60,12 @@ const createAuthenticationSession: ValidatedEventAPIGatewayProxyEvent<any> = asy
   if (isBasicAuth) {
     const auth = parseBasicAuth(authHeader);
     if (!auth) {
-      return {
-        statusCode: 401,
-        body: "Unauthorized"
-      };
+      return returnUnauthorized();
     }
 
     const elastic = getElastic(auth);
     if (!(await elastic.hasCurationsAccess())) {
-      return {
-        statusCode: 403,
-        body: "Forbidden"
-      };
+      return returnForbidden();
     }
 
     const token: string = generateToken();
